H005 Fastslå bosted: move bosituasjon fields to status, restructure Familiestatus
- Move Inntektskilde for studenter (155), Hvor permanent er bosituasjonen (155)
  and Permanent opphold skattemessig (65) from the family section into
  "Personens status og aktivitet", incl. their length validation
- Rename BostedOgFamilie -> Familiestatus (component/folder/validation); the
  card heading is now "Familiestatus" and the redundant inner sub-heading is
  removed. "Grunn for flytting" now leads the section, followed by the family
  fields
- Field sizing: descriptive fields use TextArea with a character counter
  (grunnForFlytting 155, bosted for ektefelle/barn 255); short values stay
  single-line Input (familiemedlem, skolekrets, permanent opphold)
- Fix validation bug: mainValidation passed section labels as personName,
  yielding messages like "Tekststørrelse over 155 tegn for familiestatus".
  Now passes the real personName for all H005 fastslaabosted validations,
  matching the component-level messages
- Remove now-unused label key bosituasjon-og-familiestatus
