+++ public/locales/nb/label.json
@@ -15,7 +15,6 @@
   "informasjon-fastslaa-bosted": "Informasjon for å fastslå bosted",
   "informasjon-fastslaa-bosted-beskrivelse": "Registrer adresser, status og bosituasjon som til sammen dokumenterer hvor personen er bosatt.",
   "personens-status-og-aktivitet": "Personens status og aktivitet",
-  "bosituasjon-og-familiestatus": "Bosituasjon og familiestatus",
   "bosituasjon-hvor-permanent": "Hvor permanent er bosituasjonen",
   "bosted-barn": "Bosted for barn",
   "bosted-ektefelle": "Bosted for ektefelle",
